adminpage

+++ admin/src/main/java/com/example/demo/vo/PackageVO.java
@@ -2,12 +2,13 @@ package com.example.demo.vo;
 import lombok.Data;
 
 @Data
+
 public class PackageVO {
-    private Integer packageNo;
-    private Integer type;    
-    private Integer hall;     
+    private Long packageNo;
+    private Integer type;
+    private Integer hall;
     private Integer studio;
     private Integer dress;
     private Integer makeup;
-    private Integer discount; 
+    private Integer discount;
 }
